marketplace_app/requirements.txt
asgiref==3.5.0
astroid==2.11.5
celery==5.2.7
dill==0.3.5.1
Django==3.2.13
django-author==1.2.0
django-bootstrap-modal-forms==2.2.0
django-celery-beat==2.3.0
django-cuser==2017.3.16
django-environ==0.8.1
django-formtools==2.3
django-import-export==2.8.0
django-import-export-celery==1.3
django-cuser==2017.3.16
django-environ==0.8.1
django-formtools==2.3
django-import-export==2.8.0
django-import-export-celery==1.3
django-modeltranslation==0.18.2
django-super-inlines==0.2.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.12
django-timestampable==1.0.3
future==0.18.2
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.1.0
platformdirs==2.5.2

pylint==2.13.9
pytz==2022.1
redis==4.3.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
wrapt==1.14.1